Добаивил Float  в контролере корзины

+++ catalog/controller/checkout/cart.php
@@ -280,7 +280,7 @@ class ControllerCheckoutCart extends Controller {
 
 		if ($product_info) {
 			if (isset($this->request->post['quantity'])) {
-				$quantity = (int)$this->request->post['quantity'];
+				$quantity = (float)$this->request->post['quantity'];
 			} else {
 				$quantity = 1;
 			}
